Update test for table schema
